requirements.txt
# 2025 versions - Cross-platform requirements
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
WTForms==3.1.2
Werkzeug==3.0.4
pandas==2.3.0
numpy==2.3.0
plotly==5.24.1
reportlab==4.2.2
openpyxl==3.1.5
APScheduler==3.10.4
python-dotenv==1.0.1
email-validator==2.2.0
Pillow==10.4.0

# Database drivers - platform specific
# For Windows (SQL Server)
PyMySQL==1.1.2; sys_platform == "win32"
pyodbc==5.3.0; sys_platform == "win32"

# For Linux/Deployment (PostgreSQL)
psycopg2-binary==2.9.9; sys_platform == "linux"

# For deployment platforms (Render.com, Heroku, etc.)
psycopg2-binary==2.9.9; platform_system == "Linux"